Rename infogath/ofpromq package to infogath/promq
